api/requirements.txt
aniso8601==3.0.0
bcrypt==3.1.4
blinker==1.7.0
certifi==2018.4.16
cffi==1.16.0
chardet==3.0.4
click==8.1.7
coverage==4.5.1
coveralls==1.3.0
cryptography==41.0.7
docopt==0.6.2
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
gunicorn==19.8.1
idna==2.6
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.3
nose==1.3.7
passlib==1.7.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.18
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
waitress==2.1.2
Werkzeug==3.0.1
